+++ internal/errs/exitcode.go
@@ -1,5 +1,7 @@
 package errs
 
+import "errors"
+
 type ExitCodeable interface {
 	ExitCode() int
 }
@@ -23,4 +25,19 @@ func (e *ExitCode) Unwrap() error {
 
 func (e *ExitCode) ExitCode() int {
 	return e.code
+}
+
+// ParseExitCode checks if the given error is a failure of type ExitCodeable and
+// returns the ExitCode of the process that failed with this error
+func ParseExitCode(err error) int {
+	if err == nil {
+		return 0
+	}
+
+	var eerr ExitCodeable
+	if errors.As(err, &eerr) {
+		return eerr.ExitCode()
+	}
+
+	return 1
 }
